Added strings to the strings file to account for all of the strings in the file.
Modified the TopSpot Array to take in int as an input for the referenced strings.

Modified all of the arrays to reference the appropriate strings.

Modified the arrays name to match the purpose of that specific array.

Added the pictures to the files (i think that this was the related to this commit).

Updated the list_item text to match the purpose of the application.